feat(*): Add option to normalise LR by number of graphs in batch
The sparse batching strategy means that the number of graphs per batch
can vary wildly (in VarMisuse, by factors of up to 10x).

For tasks in which the loss is defined per-graph (e.g., VarMisuse, QM9)
and not per node (e.g., PPI) and the loss is averaged across instances
in a batch, this means that the contribution per graph to gradient
updates is determined by its size and the size of other graphs in its
batches.
The introduced hyperparameter allows to define an expected average
number of graphs per batch, and then scales the learning rate with the
actual graphs in a batch.

[A nice side-effect of this is that you don't need to figure out how
to adjust the learning rate when changing max_nodes_in_batch to fit
into the GPU memory]

Experimentally, this slightly improves results on VarMisuse (by about
1%).